deprecate wlr_surface.h
